feat(llvm): support i32 for `constant` and arithmetic operations. (#921)
This PR adds support for the llvm `i32` types needed in Chacha20.
Namely, it adds: i32 constant, add, lshr, or, shl, sub, xor, zext `i8`
to `i32`, sext `i8` to `i64` and their corresponding proof and
registration into the rewriter.

Support for load and store will be added in a follow-up PR.

---------

+++ Test/Passes/InstructionSelection/RISCV64/li_invalid.mlir
@@ -2,10 +2,10 @@
 
 "builtin.module"() ({
     "func.func"()  <{function_type = () -> ()}> ({
-        %one = "llvm.mlir.constant"() <{ "value" = 1 : i32 }> : () -> i32
-        %two = "llvm.mlir.constant"() <{ "value" = 2 : i32 }> : () -> i32
-        // CHECK:      %{{.*}} = "llvm.mlir.constant"() <{"value" = 1 : i32}> : () -> i32
-        // CHECK-NEXT: %{{.*}} = "llvm.mlir.constant"() <{"value" = 2 : i32}> : () -> i32
+        %one = "llvm.mlir.constant"() <{ "value" = 1 : i16 }> : () -> i16
+        %two = "llvm.mlir.constant"() <{ "value" = 2 : i16 }> : () -> i16
+        // CHECK:      %{{.*}} = "llvm.mlir.constant"() <{"value" = 1 : i16}> : () -> i16
+        // CHECK-NEXT: %{{.*}} = "llvm.mlir.constant"() <{"value" = 2 : i16}> : () -> i16
         "func.return"() : () -> ()
     }) : () -> ()
 }) : () -> ()
